Move general actions from actions.ts
